Squashed commit of the following:

commit 0d0f2b28cd6f25b3ad851cfd2538a5c7f487ad34
Author: Katura Harvey <katharve@cisco.com>
Date:   Mon Jan 11 13:01:08 2021 -0500

    http_inspect: fix type of unit test mock function

+++ src/service_inspectors/http_inspect/test/http_transaction_test.cc
@@ -46,8 +46,8 @@ FlowData::~FlowData() = default;
 int DetectionEngine::queue_event(unsigned int, unsigned int, Actions::Type) { return 0; }
 fd_status_t File_Decomp_StopFree(fd_session_t*) { return File_Decomp_OK; }
 uint32_t str_to_hash(const uint8_t *, size_t) { return 0; }
-void FlowData::update_allocations(unsigned long) {}
-void FlowData::update_deallocations(unsigned long) {}
+void FlowData::update_allocations(size_t) {}
+void FlowData::update_deallocations(size_t) {}
 }
 
 THREAD_LOCAL PegCount HttpModule::peg_counts[PEG_COUNT_MAX] = { };
